day 7

+++ src/days/day1.rs
@@ -5,7 +5,7 @@ use std::collections::HashMap;
 pub struct Day1 {}
 
 impl Solution for Day1 {
-    fn solution(input: &str) -> (Option<i32>, Option<i32>) {
+    fn solution(input: &str) -> (Option<i64>, Option<i64>) {
         let mut heap1 = BinaryHeap::new();
         let mut heap2 = BinaryHeap::new();
         let mut count = HashMap::new();
@@ -13,8 +13,8 @@ impl Solution for Day1 {
             .lines()
             .map(|x| x.split_whitespace())
             .for_each(|mut x| {
-                heap1.push(x.next().unwrap().parse::<i32>().unwrap());
-                let right_val = x.next().unwrap().parse::<i32>().unwrap();
+                heap1.push(x.next().unwrap().parse::<i64>().unwrap());
+                let right_val = x.next().unwrap().parse::<i64>().unwrap();
                 count.insert(right_val, count.get(&right_val).unwrap_or(&0) + 1);
                 heap2.push(right_val);
             });
